ELF object files support added
- Removed manual handling of relocations
- Removed program headers creation
- Added external function declaration directives for elf backend
- Added 'VERBOSE' switch for dumps
- Moved tests to a separate folder

+++ src/transpiler/transpiler.cpp
@@ -13,7 +13,7 @@ static int get_file_sz_(const char filename[], size_t* sz)
     if(stat(filename, &buff) == -1)
         return -1;
     
-    *sz = buff.st_size;
+    *sz = (size_t) buff.st_size;
     
     return 0;
 }
